Move OasAgent::AgentContext to correct place
